:sparkles: feat: check if column belongs to board

+++ backend/src/api/column/column.service.ts
@@ -1,6 +1,7 @@
 import {
   Injectable,
   InternalServerErrorException,
+  NotFoundException,
   UnauthorizedException,
 } from '@nestjs/common';
 import { ColumnType } from '../../utils/@types';
@@ -72,6 +73,15 @@ export class ColumnService {
         'you do not have permission to perform this action',
       );
 
+    const belongsToColumn =
+      await this.boardRepository.checkIfColumnBelongsToBoard(
+        boardId,
+        columnData.columnId,
+      );
+
+    if (!belongsToColumn)
+      throw new NotFoundException('this column does not seem to exist');
+
     try {
       await this.prisma.column.update({
         where: { id: columnData.columnId },
@@ -112,6 +122,12 @@ export class ColumnService {
       throw new UnauthorizedException(
         'you do not have permission to perform this action',
       );
+
+    const belongsToColumn =
+      await this.boardRepository.checkIfColumnBelongsToBoard(boardId, columnId);
+
+    if (!belongsToColumn)
+      throw new NotFoundException('this column does not seem to exist');
 
     try {
       await this.prisma.column.delete({
